change last_fixed field in DateTakens to int

+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/models/DateTaken.kt
@@ -12,4 +12,4 @@ data class DateTaken(
         @PrimaryKey(autoGenerate = true) var id: Int?,
         @ColumnInfo(name = "full_path") var fullPath: String,
         @ColumnInfo(name = "parent_path") var parentPath: String,
-        @ColumnInfo(name = "last_fixed") var lastFixed: Long)
+        @ColumnInfo(name = "last_fixed") var lastFixed: Int)
